Update import path and version for bbolt
For some reason the bbolt maintainers started tagging their releases
with prerelease-style tags a while back. So dependency managers are
going to automatically pick up a version that's more than 2 years old.

In other news, the import path is better now and no longer tied to
GitHub, which is super neat!