Make the randomizer a bit more excited!
It's really happy that it picked these options for you! It loves picking
random options! Yay!

But in addition to making me happier I also kind of think it's easier to
read the output without the ellipses. Honestly I should probably cut
back on my use of them in general…