Make Slack package do HTTP handling only
I thought that the additional abstraction between HTTP and responses to
Slack would help ease the creation of the Lambda integration. In
reality, it turned out that figuring out how to map the model I had
created back to HTTP was also non-trivial, and duplicating that between
the normal HTTP server and the Lambda handler for their different types
of integrations was not really worth it. Much easier to write that code
once in a standard HTTP handler and make an external package map that
back to API Gateway's expected format.