Move flag parsing logic to a new file
I was looking into extending the "-n" flag to support either a count or
the string "all," but doing it in the single giant file seemed unclean.
Doing it all in a brand new file just for that functionality seemed
silly. I think that this is a decent split though.