Add initial tests for the randomizer
Now that I've started running into some actual regressions caused by me
changing this code, it's probably a good time to test it. Tests for now
are completely focused on the core randomizer component, not so much on
ancillary functionality like the various store implementations (which
might not be trivial to test anyways).