Unit test responses on transient store errors
Notably, this revealed a case where we do not distinguish between a
group not existing and being unable to get a group due to an unexpected
error. Need to look into this more, probably by using nil as a sentinel
for a group that we can say with certainty does not exist. Would have to
figure out what that looks like in DynamoDB and Bolt though.