Rename validIf to check in tests
This is a bit more natural to read, both in the test definitions and
especially the test driver.

I also snuck in a change to the formatting of the error when store
expectations are not met. Now, both stores are displayed at the same "x"
position in the terminal, so they're easier to compare.