Allow "slashy" flags in arguments
This is an attempt to separate the character used to remove items (the
minus sign) from the character that represents flags (now a slash), so
that flags can appear anywhere in the command.

This is just a basic implementation to allow the tests to be updated.
New flag parsing logic will be used so that slashes are the *only*
allowed character for the arguments.