Implement experimental version of slashy options
This is the "full" version of slashy options, where *only* slashy
options are allowed, and they can appear anywhere in the command string.

This is a really crappy implementation just to start things off. If it
seems like the new style will be worthwhile to keep around, the code can
be cleaned up.