Write a new slashy flag parser from scratch
In the long run, this is going to be a slightly less terrible way to do
slashy flag parsing. Rather than maintain the patterns established by
prior use of the built-in flag package, this rewrite introduces a new
type representing parsed options specific to the randomizer.

Note that there are a bunch of remaining special cases and hacks that
need to be handled better. This is an ongoing work in progress.