Implement more test coverage for slashy parser
In particular, implement some test cases that put the "/n" flag at the
end, which is new functionality.

Also add some tests that provide "/show" and "/delete" flags but omit
the required group names, and a test for "/save" that omits the things
to save, to try out some edge cases in the parser.