Add tests for saving expanded group options
This started out half as desired functionality, and half as a side
effect of the gigantic, crazy App.Main method. It's sort of
undocumented, but I do think it's useful and I want to keep it around.
Previously the expansion behavior was implicitly tested through some of
the other tests, but with the split-out of Main I've added some extra
examples to introduce coverage to this particular use case.