Separate missing groups from data store errors
Previously, attempts to get a non-existent group resulted in a data
store error. While this makes some sense, it prevented the randomizer
from distinguishing between a group not existing and a more generic data
store error. In the former case we want to tell the user to create the
group (or maybe check their spelling), but in the latter case we want
the user to know that they should just try again later.

With this change, non-existence of a group is now represented by
successfully returning an empty list (similar to how the List method
already worked), while errors are only used to represent (potentially
transient) data store failures.

As part of this, documentation was added to Store interface methods to
better communicate their expected behavior. It was noted that the Delete
method is a bit problematic right now, as it can't indicate whether the
deleted group actually existed prior to deletion. This will be fixed in
the future.