Remove time bomb "continue" from flag parser
Two things:

First, the parser works fine without this continue statement. It's not
really any less efficient, in fact it's a bit more efficient since we
can start handling non-modifier arguments in the same loop iteration
instead of doing an extra unneeded length check.

Second, if I ever need a case where a modifier handler consumes 0
arguments, this statement would have created an infinite loop, as the
non-modifier would never go on to be consumed as a regular argument.