Stop trying to check for mistyped flags
If a user saves a group with a name starting with a slash that does not
conflict with an operation, this logic would prevent that group from
being used. Which may or may not be a bad thing. If it's ever confusing
enough for people I might add this back in.