Update DynamoDB table provisioning settings
Two separate changes here.

First, the CloudFormation template has been updated to use on-demand
capacity mode, because HECK YES.

Second, the default provisioned capacity for the randomizer-dbtools
creation logic was changed from 10 to 1 read / write capacity unit(s)
each. This is to make things slightly cheaper in the default use case.
I'd have updated this to use on-demand capacity as well (or at least
made that an option), but the (beta) AWS SDK v2 for Go hasn't been
updated with that feature yet (there's a GitHub issue that says they're
trying to catch up with all the re:Invent announcements in general).