Use strings.Fields to split Slack command text
I'm honestly kind of ashamed that I never knew about this before. Like,
this solves a lot of problems for me. In this particular case, it should
help solve the problem where running the randomizer with an extra space
at the end but no group name leads to an error.