Support BuildKit-powered cross-compilation
This is sort of a proof-of-concept Dockerfile update to test out
multi-platform image builds using BuildKit's QEMU / binfmt_misc support,
using Go's native cross-compilation support to avoid the overhead of
virtualizing the compiler.

The fact that the OCI specification literally defines platform names in
terms of GOOS and GOARCH makes this surprisingly easy.