Apply context to randomizer operations
This should be enough to start getting us X-Ray traces for DynamoDB
operations on AWS Lambda.

I decided to shove the randomizer context into the "request" struct that
everything already relies on, since it seemed like it fit best into the
existing model compared to changing the signatures on all the functions.