Set up deployments to use a container image
We now have the Lambda function using a container image instead of a zip
archive!

Note that the current configuration does NOT work right, since there are
no TLS roots in the image. We can randomize lists of things, but we
can't get to DynamoDB to do anything with saved groups.

I had to change the name of the resource in the CloudFormation template
in order to create a new Lambda function. It turns out Lambda does not
let you update from an archive-based function to an image-based
function; you have to create a brand new function and re-link everything
to it.