Add TODO about SlackToken parameter
This feels like a secret, right? Like, something that should NOT just be
a stack parameter and a plaintext Lambda environment variable? Like it
should be a KMS-encrypted SSM parameter?

How have I never noticed this before?

Anyways, I think I'll have to pass the full KMS key ARN as a stack
parameter to avoid giving the Lambda function too broad of a kms:Decrypt
permission. If this were Terraform I'd just run some data lookups on the
SSM parameter to find the KMS alias and then look up the full key ARN
but... you know. Maybe cf.sh can do a bunch of AWS CLI crap to find that
value itself?

I started looking into CDK, but it can't import CloudFormation templates
that rely on transforms, so I'd have to rewrite the whole stack from
scratch to even experiment with it properly. At least I'd be able to do
all of that lookup junk, though. In TypeScript, even!