Support Slack token from SSM in randomizer-server
Similar to the unification of the store factory creation logic across
different randomizer binaries, the Lambda and HTTP servers now both
support the same environment variables to configure the Slack token.
This way, if you run the HTTP server in AWS or otherwise have AWS
credentials for it, you can optionally leverage SSM for the verification
token instead of plainly making it available in the environment.