Remove API gateway support from the CloudFormation stack
**THIS IS A BREAKING CHANGE!!!**

I've added warnings to the documentation about how this will destroy
existing URLs, and about how to migrate smoothly to the new setup, just
in case. However, according to GitHub it seems like there are literally
zero other people who care about this repo at all, so if someone is
actually using this and doesn't see the message I guess that's just
going to be the consequence of not adding a star or becoming a watcher.