Configure memory-related settings more explicitly
GOMEMLIMIT will apply to the randomizer when built with Go 1.19 or
above, and will be ignored otherwise. Honestly, no real randomizer
deployment should ever come close to this limit, so think of it as more
of a general best practice, or a life philosophy where we ask "why not?"
instead of "why?", which some questionable Internet quote sites tell me
is something JFK once said.