Rewrite the argument parser
In full transparency, I wrote the first version of this parser in part
to try out the so-called "Clean Code" philosophy of writing bunches and
bunches of small methods. That approach felt vaguely "off" to me then,
and I've long since come to *know* that it's a bad approach after
finding multiple great resources that put the issues into words better
than I ever could [*], and I finally got sick enough of looking at it to
want to rewrite it.

I have slightly mixed feelings about the structure of the main switch
statement in this new version (with some cases having return paths and
others falling through past the switch body), but I still like it better
and find it easier to understand than the old function-based approach.

[*]: The most helpful explanation was the "Modules Should Be Deep"
chapter of "A Philosophy of Software Design." Later on I came across
https://qntm.org/clean, which focuses even more explicitly on the "small
methods" aspect (and, in fact, also recommends "A Philosophy of Software
Design").