Simplify the list formatting routines
Both of the routines now use a fairly similar strings.Builder
implementation to build lists out of multiple elements. The routine for
inline lists of bold items is especially a lot easier, since instead of
two functions that were called together 100% of the time we just have a
single string build.

I also threw strings.Builder into the help implementation as a
replacement for bytes.Buffer, since it's just as easy to use for our
particular case.