Clean up documentation throughout the codebase
I've also consistently put package comments everywhere so that I can
enable the package comment rule in staticcheck. Note that I'm NOT just
enabling this mindlessly so that I can have "all the rules." I think
that these short summaries are legitimately useful when browsing
documentation for the module (as evidenced by the fact that several
packages *did* have them, as well as the fact that some of my other
projects include them more consistently), which is why I consider their
omission to be an accident, and why I have explicitly and intentionally
chosen to enable this specific rule and enforce that future packages
also meet this standard.

(On the other hand, I think I deserve to be called out for caring enough
to tweak specific rules, but not enough to enforce staticcheck on every
push to the repository. If I go that route, I probably need to enable
and justify each individual rule to prevent future staticcheck changes
from unexpectedly breaking builds. I currently use staticcheck on a
best-effort basis during local development only, which means that it's
fine to have something break and start showing up as a warning in my
editor. But by that logic, I SHOULD have set this to "all" in the first
place and allowed those warnings to remain in my editor until I chose to
fix them, instead of temporarily hiding them from my editor with a
config change.)