Migrate all Slack error logging to log/slog
This includes a full update to the Lambda handler, which doesn't
actually do much logging of its own.

Within the Slack handler, I'm doing something vaguely like the logr
wrapper that Kubernetes uses internally, with the error followed by a
message.

Technically it would be smart to include the request context in my logs,
but I don't think it's that big of a deal.