Take "%+v" error formatting out of randomizer-demo
The move to slog eliminated nearly all of the formatting strings for the
stack traces that github.com/pkg/errors adds to its error values. It's
probably time to accept that stack traces in errors haven't received the
kind of reception I'd initially hoped for (for reasons that are
admittedly valid in a lot of contexts), and flip back to the native
errors package across the board.