Migrate entirely to native Go errors
This removes all direct use of github.com/pkg/errors, leaving it as a
dependency of the AWS X-Ray SDK.

This was a lot trickier than I expected! At first I forgot about the
semantic of errors.Wrap[f] where it returns nil if the original error is
nil, and I nearly introduced a whole bunch of new errors where I
shouldn't have. In fact, even after reviewing the code about 5 times,
I'm still worried I've missed something. Sadly, my testing around error
cases is not great, especially for the lower levels of the DynamoDB and
bbolt stores. I'll just try to be as comprehensive as possible in manual
testing before I ship this anywhere important.

On the plus side, this was a great opportunity to clean up a lot of
weird patterns, like calling an important method with important side
effects in the argument list of a big errors.Wrap call. I also found a
case where I was wrapping an error twice with the same message, even
though the first of those wrappings was already in an "if err != nil"
check, such that the second wrapping would always be a no-op.