Stub out a Firestore client
I'm experimenting with Cloud Run, so I need a place to store data that
isn't a filesystem (and is ideally cheaper and easier to operate than
running GKE Autopilot just to use bbolt on a persistent volume).

This "only" adds about 5 MB to the randomizer-demo binary, which I guess
isn't quite as bad as I feared. We'll see what happens when I start
throwing real documents around, though.