Put Firestore support behind a build tag
I'm going to treat this as experimental for now, at the very least
because it has no documentation.

With this change, the randomizer binaries without the build tag are back
to the same size that they are on the main branch, which is a little
weird since I would have expected the root FactoryFromEnv integration to
have at least some effect.