Fix up how randomizer-server configures signals
I typo-ed the name of the slice containing the signals on non-Unix
builds, so this wouldn't have compiled at all on those systems. Hardly
the first time I've written something Unix-only that "seems" like it
could be cross-platform. Should I bother with cross-platform GitHub
Actions testing? Not sure if it's worth it for this specific project.