Return HTTP 405 + Allow for invalid HTTP methods
…and write a comment about, um, how WILDLY F@%#ING INCOMPETENT I AM for
implementing a web service with non-idempotent / non-safe /
non-whatever-the-official-terminology-in-the-RFC-is HTTP GET requests.

How has it taken nearly 6 years to see what DEFECTIVE GARBAGE this is?

WHAT IS **WRONG** WITH ME?!?!?

This is a breaking change that MUST be made. I'm almost certain that all
of the integrations I've personally set up are using HTTP POST, because
thankfully I've had at least one moment in my life where I didn't
actively pollute the world with this kind of toxicity.