Remove fundamentally invalid HTTP GET support
For legacy reasons, the original version of the slash command handler
supported both GET and POST requests to the same endpoint, matching the
two methods Slack supported for the legacy slash command integration.

Two things have happened since then.

First, after revisiting the HTTP semantics of this handler more
meaningfully, it became clear to me that the support for HTTP GET
requests was fundamentally BROKEN, DEFECTIVE, GARBAGE, and WRONG.
Because the randomizer can easily perform non-idempotent actions based
on the provided command text, it can NEVER claim to implement proper
HTTP GET (and HEAD) semantics (at least without significant
implementation changes, i.e. the removal of persistent storage).

Second, Slack has deprecated the GET option when adding slash commands
to apps, and only supports integration through HTTP POST.

Given that previous references to the option of GET setup were deleted
in a rewrite of the documentation, and that all of my current randomizer
usage is with the POST form of the integration, there is absolutely no
justification to maintain this fundamentally BROKEN feature that was the
product of DEFECTIVE and UNACCEPTABLE thinking.