Bump dependencies
It's been about a month since, uh, whatever that last commit was where I
went wild with hot takes. So here we are again, but keeping things more
chill this time. After all, "wouldn't it be tight if everyone was chill
to each other?"

The big story of this one seems to be go.opentelemetry.io/otel/semconv,
which I guess standardizes how people emit their OTel traces?

I'm starting to wonder about my commitment to Google Cloud support. It
was a decent learning experience, but my tooling to manage Cloud Run
isn't as built-out as what I have for CloudFormation, and it yields a
big dependency tree.