Bump all dependencies
Wow! So, it turns out that Slack's deprecation is only for legacy custom
_bots_, and not the legacy slash command integration (which some
instances of this system definitely still use).

The most interesting thing about this is that bbolt is receiving
updates! v1.4.0 came out yesterday, and for the most part it looks like
a lot of internal refactoring. This version has been in a prerelease
state for over a year, so I'll assume the etcd folks have worked out any
issues by now.

(And yes, I read about the typosquatting attack on an unrelated fork of
Bolt, for whatever that's worth.)

Once again the binary protobuf stuff annoys me. Protobuf-related code
feels like quite a substantial portion of these changes. But I'm still
just complaining about it without taking action, so...