Remove an unused build stage (and add a note)
The question posed in the previous commit was a legitimate question on
my part, but I failed to make that clear and instead just wrote it
without context while I pushed it and went to search for the answer.
And on second reading, I feel like it comes across as having an
accusatory tone toward the maintainers of the official open source
Docker image, which is actively harmful of me.

I want to append onto that commit to note: I expected that there would
be a clear explanation for this default, and there was, and it is clear
that the correct default was set for this context. More importantly, I
was able to confirm that I'm not doing something _truly_ horrible, i.e.
that this wasn't set because it would actively break something around
how the Go distribution is configured in this specific image.

I apologize for the poor message on my part, and hope that this
additional note serves as a concrete step in the right direction.