Remove internal/slack dependency on internal/store
Because the Slack package imported the entire store package for the sake
of a single defined type that wasn't even used anywhere else, the Lambda
handler was getting a bunch of Firestore junk linked in that I had
clearly not intended (as evidenced by me importing only the DynamoDB
store into the main package).

This shaves about 4 MB off of the darwin/arm64 copy of the binary, which
admittedly isn't a great metric to base this on but is the easiest one
for me to look at.

There's still work to be done here, since I'm still seeing the gRPC
libraries linked in even though Firestore is gone. Is Amazon secretly
using gRPC too? Or have I missed some other dependency?