Add a test for banned packages in randomizer-lambda
This is to help keep me honest about my recent optimization work, and is
again influenced by the fact that the Cobra change included a test.

Note that I don't just match MethodByName like that test does. We have
some MethodByName calls in the final binary, which are fine because the
compiler can statically determine the methods being looked up. It's only
when you call that method with a dynamic value that you block all
possibility of dead code elimination.