Rework store construction based on the new registry
I don't like this, though, because if you set (for example) DynamoDB
environment variables and it's not linked into the binary you'll
silently fall back to a bbolt store. At the same time, I don't want to
break the bbolt fallback or force it to be linked in every time.

I'm thinking about defining all of the environment lists in the registry
up front, and only registering constructors from the store packages.