Report an error when asking for an unlinked backend
This solves the user configuration concern by intentionally taking on
the risk of a store backend being defined incorrectly.

Of course, I don't love this approach either. My long-term plan is to
have build tags for each store, and a file that links in all stores if
none of the build tags are set. I'm now thinking that file should set a
flag to indicate that all possible stores are available, so it's safe to
fall back to bbolt. If some stores are missing, the bbolt fallback is no
longer safe since the configuration for an unlinked backend could
trigger it.