Set up build tags and new bbolt fallback logic
Now, if we have all stores guaranteed to be available, or if bbolt is
the only store linked into the binary, we'll select it as the fallback
if no other environment variable configuration is found. Otherwise, we
might simply have missed the environment settings for a non-linked
store, so we can't safely fall back.