Note an alternative approach to bbolt-only fallback
Not sure why I didn't think of this the first time, rather than the wild
approach I took with all the init functions. Maybe I was hesitant to not
follow the common error handling path, even though it's not relevant.

I still don't want to do this, though. The code is simpler, but the
concept is still hard to explain.